+++ app/Http/Controllers/Task/TaskController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Project\UpdateProjectRequest;
 use App\Http\Requests\Task\StoreTaskRequest;
 use App\Http\Requests\Task\UpdateTaskRequest;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use App\Models\Task;
 
@@ -16,6 +17,7 @@ class TaskController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Task::class);
         return view('control-panel.tasks.index',['tasks' => Task::paginate(10)]);
     }
 
@@ -24,6 +26,7 @@ class TaskController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Task::class);
         return view('control-panel.tasks.create');
     }
 
@@ -32,6 +35,7 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
+        $this->authorize('create', Task::class);
         $task = Task::create($request->validated());
 
         return to_route('tasks.show', $task);
@@ -42,6 +46,7 @@ class TaskController extends Controller
      */
     public function show(Task $task)
     {
+        $this->authorize('view', Task::class);
         return view('control-panel.tasks.show', ['task' => $task]);
     }
 
@@ -50,6 +55,7 @@ class TaskController extends Controller
      */
     public function edit(Task $task)
     {
+        $this->authorize('update', Task::class);
         return view('control-panel.tasks.edit', ['task' => $task]);
     }
 
@@ -58,6 +64,7 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        $this->authorize('update', Task::class);
         $task->update($request->validated());
 
         return to_route('tasks.show', $task);
@@ -68,6 +75,7 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        $this->authorize('delete', Task::class);
         $task->delete();
 
         return to_route('tasks.index');
